Cancel broadcast context before signature
When member jumps outside the message loop, it
has enough valid shares and can complete the
signature in the next step. Because of
that, share exchange context can be cancelled
at this stage because there is no more sense to
continue the communication with other members.
This way we can save some resources and eliminate
useless message retransmissions.