Check current entry status after signature error
In case a threshold signer returns with error,
the current entry status should be checked.
If current entry is not in progress anymore,
it means the result was already submitted by an
other signed and the error should not be logged.
If the current entry is still in progress,
something went wrong and signature error
should be logged as usual.