Conservative relay entry timeout punishment for the first months
We dediced to use a conservative relay entry timeout punishment for the
first months of the network so that we can give time to operators to
properly configure their clients and learn how to monitor their health.

The rules are following:
- 1% of the minimum stake for the first 3 months
- 50% of the minimum stake between the first 3 and 6 months
- 100% of the minimum stake after the first 6 months.

This change affects only relay entry timeout, all the other cases of
misbehavior are still on 100% of the minimum stake.