Do not cancel the context before entry is on-chain
We added cancelCtx() call to cancel signature shares retransmission if
the current member has enough shares to reconstruct signature. From
a selfish perspective of that member it absolutely makes sense but from
a broader perspective it does not and it slows down relay entry
submission.

Specifically, if members 10-64 collected enough shares to reconstruct
entry, they stop broadcasting. If members 1-9 still do not have enough
shares, none of them will be able to publish the result and we'll have
to wait until member 10 is eligible.

Retransmitting shares until the entry is on-chain makes sense.