Use groupSelectionStarterContract to reimburse for failed DKG
In startGroupSelection we return the costs of the previous - failed -
group selection to the service contract. We were obtaining the service
contract reference with `ServiceContract(msg.sender)`. It works fine
only when `msg.sender` is a service contract which is not a case for
genesis. Instead of `msg.sender` we should use
`groupSelectionStarterContract` that is set for this exact purpose in
`genesis` and `createGroup` both calling the internal
`startGroupSelection`.