Improved relay request confirmation mechanism by adding retries
The confirmation mechanism now has built-in retries. We can retry in
case of an error but also when the expected request start block does not
match the one currently registered on the chain. Such situation may
happen for Infura-like setup when two or more chain clients are behind a
load balancer and they do not have their state in sync yet.