Log getting selected participant retries on info level
Experience shows that this situation happens very often on Infura-like
deployments and ERRORs produced there were adding too much noise given
that usually the second attempt was successful. We do not want to
clutter logs with ERRORs that are not really errors - we need to accept
that it's how most ETH clients load balancers works.