Pass compressed public key to archive function, improved error handling
When we save membership in the storage, we execute
GroupPublicKeyBytesCompressed() on the membership passed as a parameter,
so that we can evaluate name of the directory where we want to put the
membership.

Groups registry, on the other hand, holds group public key in an
uncompressed form so that memberships can be quickly fetched from the
registry when on-chain events occur.

When the group registry was asking the storage to archive group, it was
passing uncompressed public key and the storage could not locate the
directory to archive.

It is fixed here - storage expects uncompressed public key in archive
function. I have also updated error handling in the UnregsterStaleGroup
loop. We now explicitly skip the rest of the step execution in case of
an error. The previous version worked just fine because isStaleGroup is
false on error, but having the explicit continues instructions is
cleaner and less error-prone.