Cleanup script
Here we wrap migration dir cleanup in its own script for execution in a
separate run command on Circle.  This, coupled with the config parameter
`when: always` will let us clean up the utilitybox migration dir whether
the migration job passes or fails.  As-is it cleans only when the job
passes, this results in the utilitybox disk getting filled up when there
are a few failed builds.