Message dissemination is disabled by default and can be configured
We introduced message dissemination to handle cases when nodes from the
same topic are not connected with each other and topic messages are not
relayed between them. It helped with message delivery but also caused
problems with throttled message validation on slow clients.

Message dissemination is enabled only for relay entries and given that:
- we have now 6 * 64 blocks to submit entry before the timeout,
- client tries to connect to all bootstrap nodes from the list,
- in most cases, dissemination is not needed and it's just a last-resort
option when something goes wrong with the network,

we should enable it only on selected bootstrap nodes. All clients tries
to connect to all bootstrap nodes so dissemination on bootstrap nodes
should be enough. Even more, since bootstrap nodes also try to form a
dense mesh with all other bootstrap nodes, it is fine to enable
forwarder on just some part of bootstrap nodes.

Here we disable forwarder (dissemination) by default and let to enable
it in the config allowing to parametrize the dissemination timeout. This
should help with fine-tuning the network for early days based on the
load and number of clients.