Use keep-common nonce manager in all contracts
Nonce manager is used by the generated contracts to track the current
nonce and update it after each successfully submitted transaction.
Tracking the nonce locally is required when transactions are submitted
from multiple goroutines for when Ethereum clients are deployed behind a
load balancer and their mempools are not always in sync.

In our case, we are having problems with ticket submission against both
Infura and Alchemy, and we have a feeling are load-balanced Ethereum
APIs are affected. To be 100% safe, we need to handle nonce on our side.