Don't reflect current operator in authorizations for non-operators
In cases where no operators were found for the current address as if the
current address was an owner, operator lookup assumed the current address was
an operator. This was not always true, and allowed adding available bonded ETH
to an address that had no delegation and therefore had no way to recover it. We
now check that the current address is an operator (by checking for its owner)
before treating it as such.