Removed p == 2 condition in modSqrt Solidity implementation
p == 2 is not a valid input. P has to be an odd prime and this
requirement is made clear in the documentation of the function.

Given that modSqrt is an internal function and p value is hardcoded in
AltBn128 library, we should be fine with no additional check saving some
gas on relay entry verification.