Library determining grantee and token owner
It is not always easy to determine who is the "owner" in the stake
delegation "owner-operator-beneficiary" trio. Specifically, for a
scenario when the delegating party owns tokens, the owner is just the
token owner address. However when delegated tokens come from a grant,
the owner of tokens is a TokenGrant contract but owner in the delegation
trio is grantee. This is even more complex for ManagedGrant, because in
this case, the owner in delegation trio is the grantee from ManagedGrant
contract.

Being able to determine who is the owner is important for use cases like
KeepBonding contract where we need to figure out who can make a decision
about withdrawing unbonded value.