Added nonReentrant modifier to submitDkgResult
Let's consider the following:
- It is legal to stake an operator that is also a smart contract
- A smart contract is allowed to submit tickets, as long as it
  generates the proof of work correctly
- The beneficiary can be equal to the operator

So, stake a smart contract and set the beneficiary to the same
address. Also stake an account you plan to use. When the time
comes, submit some tickets. If both accounts win, proceed normally
with the normal account but don't bother providing a signature for
the contract.

When your smart contract makes the call to submitDkgResult (which,
hopefully you optimized for it to be first or close to first), the
beacon operator will perform a callback to your smart contract.
You can keep calling submitDkgResult as many times as gas limits
allow, each time adding a "new" group to the system (they'll all
share the same public key). This gives your group a much higher
chance of being selected to generate a random number, which is to
some degree an advantage.

Adding nonReentrant modifier to submitDkgResult should prevent this
scenario from happening.