Do not allow to slash group more than one time for entry timeout
When a group is inactive, they can be slashed via
`reportRelayEntryTimeout`. If there is another group, then the signing
process will be restarted. However, if there is only one group in the
system we can't restart the signing process and relay request remains as
timed out.

we had a situation when because of a lack of a validation, and only one
group exsting in the system, that one group could be slashed multiple
times in case of inactivity.

I refactored Groups library adding terminated status check to the
terminateGrop function. This way, we can be sure the transaction is
going to revert for already terminated function. Also, it allowed to
simplify the code of `reportUnauthorizedSigning` function where we can
now return a preceise revert message.

Added two unit tests making sure relay entry timeout/unauthorized
signing can not be reported more than one time for the same group.

One additional change that could be implemented here was resetting
`currentRequestStartBlock` in `reportRelayEntryTimeout` to zero.
I decided for not doing that given that:
- If there are groups in the system, `currentRequestStartBlock` will be
set to the new value when restarting signing process.
- If there are no groups in the system setting
`currentRequestStartBlock` does not help because we need to genesis
again and new relay request can be done later anyway if the previous
entry timed out.