Prepare Groups library and operator contract for rewards
Expose publicly the total number of groups created by the operator contract,
and the registration blockheight and termination status of a group.
To enable this, add a function in Groups to query the registration blockheight.
To prevent out of gas issues on deployment,
change `setGroupMembers` to public instead of internal.