Owner of TokenStakingEscrow is TokenStaking
TokenStakingEscrow is made Ownable to eliminate chicken and egg problem
between TokenStaking and TokenStakingEscrow. TokenStaking can't create
TokenStakingEscrow in the constructor because it fails with OOG. It
means that we need to create TokenStakingEscrow first and pass it to
TokenStaking. However, since TokenStakingEscrow allows only TokenStaking
to create deposits we were having ordering problems. Making
TokenStakingEscrow Ownable and transferring the ownership to
TokenStaking once it's created solves this problem.