Pass TokenStakingEscrow reference to TokenStaking
TokenStaking can not work without TokenStakingEscrow so we are passing
the reference in the constructor. My initial approach was to create
TokenStakingEscrow inside TokenStaking but although it's probably the
most elegant solution, it fails with out-of-gas error.

I have also removed KeepToken parameter from initContracts function.
KeepToken is never mocked and we always use the real contract for tests.