Attempt to set ETH host for token-dashboard
Today the token-dashboard reads a config file checked into source
control to select which ETH host to use.  In an effort to obfuscate the
hosts we're going to try out crafting the read config file based on
Circle Context, during the build process.

This will likely take a bit of fiddling to get right.