Updated groupActiveTime value in stub after fixing test stake delegation
We were not returning a promise from stakeDelegate function (fixed in
the previous commit) and as a result, all await instructions for
stakeDelegate functions were useless. Now we are actually waiting for
stakeDelegate to complete in tests and because of that I had to correct
groupActiveTime in test to include the time for 2 additional awaits when
setting up a test.