Use default config
Running into an issue on the dashboard where it's telling me to select
hte correct network.  I haven't figured out how to find the config on
the container so I'm reverting to the checked in config to see if we
still get this error.

I've gone back and pulled in the last keep-core package that was
migrated against keep-dev.