Allow the grant manager to withdraw deposit for revoked grant
Added withdrawRevoked function allowing to withdraw the entire amount
that is still deposited in TokenStakingEscrow in case the grant has been
revoked. Anyone can call this function and the entire amount is
transferred back to the grant manager.

TokenStakingEscrow does not implement unlocking scheme for revoked
grants and always returns the entire, not-yet-withdrawn amount to the
grant manager. The reason for that is that it's very hard (if possible
at all) to calculate the amount that should belong to the grantee of a
revoked grant and that still remains in the escrow. We leave this to the
grant manager.