Missed dependency updates for tbtc keep-ecdsa
There is a more general issue here in that these dependencies can't be
resolved until keep-ecdsa and tbtc contract migrations are done.
keep-ecdsa and tbtc contract migrations can't be done until keep-core is
done, and the dashboard build is part of the keep-core workflow in
Circle.

This is a product of pre-tbtc dashboard integration, where the only
thing the dashboard depended on was keep-core.