Retry once on nonce too low
During periods of heavy use, namely right after a new release is
deployed to Ropsten, we run into issues with nonces being too low.  We
can combat this a bit by reducing max parallel invocations to 1 but that
slows things down quite a bit.

Here we introduce a single retry  if the initial grant transaction
returns an error  with nonce too low.  This isn't a full cure, but will
hopefully reduce the number of times someone has to slam the faucet
during heavy use.