Fix fetching tbtc rewards
Match `Transfer` events from the `TBTCToken` contract and `Created`
events from the `TBTCSystem` contract. A `Created` event contains
a deposit contract address and we match this address with `from` field
form the `Transfer` event. In that case we are sure that a `Transfer`
event is connected with a reward transfer.