Allow grantee to cancel/undelegate tokens redelegated from escrow
TokenStake needs to know from which grant tokens are redelegated.
Redelegation goes to another operator that has not been used yet
and when receiving tokens in receiveApproval, asking TokenGrant about
grant ID for the operator does not work. TokenGrant is unaware of that
operator. Thus, TokenStakingEscrow needs to inform TokenStaking about
the grant ID of tokens used for the redelegation. It's now done by
appending grant ID to extraData parameter with beneficiary, operator,
and authorizer.

Secondly, I simplified how GrantStaking.canUndelegate works as well as
simplified RolesLookup.isManagedGranteeForOperatorAndGrant - it is now
RolesLookup.isManagedGranteeForGrant. Previously, we were asking
TokenGrant to lookup in granteesToOperators to say if the given operator
has tokens delegated from the given grantee. It does not work for tokens
redelegated from TokenStakingEscrow because - again - the new operator
is not known by TokenGrant. I had to simplify canUndelegate check by just
looking up grantee by grant ID, and similarly for isManagedGranteeForGrant.
This is simpler, cheaper, and still safe because TokenStaking captures
grant ID during the first delegation and it does not allow to override it
later. Grant ID is obtained from the information held in TokenStaking
(GrantStaking library to be precise) so looking up grantee for a grant with
the given ID is safe and all we need for canUndelegate check.