Verify delegated authority in release lock
Authority of operator contract can be delegated. It happens in case of
BondedECDSAKeep contract which receives authorization from
BondedECDSAKeepFactory. In such case verification of authorization for
the operator contract in releaseExpiredLock and isStakeLocked produced
incorrect results.

We replaced call to `registry.isApprovedOperatorContract` with a call to
AuthorityVerifier contract that is implemented by TokenStaking and holds
all information regarding authority delegations.