Stake top-ups
The idea behind a top-up is to let the staker add KEEP tokens to
existing delegations. This is a two-step process: first, staker
initiates a top-up, locking new KEEPs in the staking contract.
Once the initialization period is over, staker (or anyone else)
can commit the top-up, increasing operator's stake.

From the smart contract's perspective, top-ups are performed using
the same interface as delegations: `receiveApproval` on `TokenStaking`
contract. If the given operator is active, tokens are treated as
a top-up. If the given operator is not known, tokens are used to
create a new delegation.

Operators who have their stake undelegated cannot have their stake
topped-up. It means that we still enforce operator's uniqueness -
operator address cannot be reused after delegating and recovering
the stake. This is important because token staking escrow assumes
operator's uniqueness when accepting token deposit.