Move registration blockheight into its own storage slot
When the registration block is stored in the same slot as other data,
every access of the registration block inlines some additional bit operations
to clear the excess bits.
By moving the `terminated` and `registrationTime` slots to their own fields,
we can save enough gas to avoid out of gas issues in stub deployments.