Undelegation period made a constant
Initialization period varies per network but undelegation period was
always the same. For this reason, we can make it a constant and save 16
bytes of bytecode (24431->24415) in TokenStaking. Every single byte is
precious now, when we want to add stake bridge code.

I set the undelegation period to ~2 months, the same value as we use on
mainnet.