Allow bridging already-staked tokens on a new staking contract
We'll be leveraging the tBTC deposit pause to deploy a new staking
contract with a few features that weren't possible in the v1 release
of the Keep core contracts. In particular, this staking contract
will support “top-ups”, meaning the ability to delegate additional
tokens to an existing delegation relationship (i.e., an existing
operator/beneficiary/authorizer triplet). This introduces a key
complication, however: tokens already delegated through the current
staking contract must remain locked in that contract for 60 days,
the length of the undelegation period. This means that existing
token owners who have used the old staking contract must wait 60
days to port those balances over to the new contract.

The core goal is for the Keep team to make an amount of KEEP tokens
available to temporarily stake on behalf of a token owner, provided
that they have those same tokens currently locked up in the old
staking contract. Undelegation of all tokens in the new staking
contract would be blocked until the temporarily-staked tokens
provided by the Keep team are repaid by the token owner.

The expected mode of usage for users who have already staked on the
old contract is to:
1. Copy stake.
2. Undelegate from the old staking contract.
3. Operate as normal on the new staking contract.
4. Once the old staking contract's undelegated balance is available
for recovery, recover it and use the recovered tokens to repay the
StakingPortBacker contract.