Set undelegation period to 2 weeks for the first 2 months
For the first 2 months after the deployment of TokenStaking contract,
undelegation period is 2 weeks and it goes back to 2 months once 2
months after the deployment passes.

The staking contract guarantees that an undelegated operator’s stakes
will stay locked for a period of time after undelegation, and thus
available as collateral for any work the operator is engaged in.

Two weeks is the absolute minimum given that the beacon group active
time is also two weeks. Changing undelegation period does not affect
keep-ecdsa deposits since the stake is locked by the sanctioned
application until all deposits the given staker has are redeemed and
keeps are closed.