Increased gas price ceiling to 60 Gwei
The original proposition (previous commit) was to increase the gas price
ceiling to 50 Gwei but given the current Ethereum trends, we think it is
a good idea to increase the ceiling even further, to 60 Gwei.

If we go from 30 Gwei to 60 Gwei ceiling we'll increase base (no callback)
entry price by 12.3%: 0.072982 ETH -> 0.081964 ETH and we'll increase the
fee for opening a keep by 13.73%: 0.07418878 ETH -> 0.08437756.

For 60 Gwei ceiling, entry fee is 81964000000000000 wei = 0.081964 ETH
For 50 Gwei ceiling, entry fee is 78970000000000000 wei = 0.07897 ETH
For 40 Gwei ceiling, entry fee is 75976000000000000 wei = 0.075976 ETH
For 30 Gwei ceiling, entry fee is 72982000000000000 wei = 0.072982 ETH

The impact on DKG frequency:
60 Gwei gas price ceiling, 20 Gwei actual gas price: DKG is triggered every 42 entries.
60 Gwei gas price ceiling, 10 Gwei actual gas price: DKG is triggered every 27 entries.
60 Gwei gas price ceiling, 1 Gwei actual gas price: DKG is triggered every 13 entries.