Fetching data from TokenStakingEscrow
If we fetch grant details we must to take into account deposits for a
provided grant from `TokenStakingEscrow` contract. The
recovered/canceled tokens from a grant delegation are deposited in the
`TokenStakingEscrow` and a staked value in a `Grant` struct is not
updated. To calculate the grant data correctly we have to make
additional requests to a` TokenStakingEscrow` contract.