Withdraw tokens from TokenStakingEscrow deposit
If the user recovers/cancels delegation from a grant, the user is not
able to withdraw all tokens from a given grant via `TokenGrant`. The delegated tokens
are deposited in the `TokenStakingEscrow` contract. To withdraw all
tokens the user must call a `withdraw` fn from a `TokenGrant` and
`TokenStakingEscrow` contract