Recognize grant delegation correctly for multiple staking contracts
We can have a situation when someone staked to operator A from a
TokenGrant using another staking contract. In this situation, current
staking contract may or may not consider the delegation as coming from a
grant. To make the right decision, the current staking contract should
check if the staking contract address returned from
TokenGrant.getGrantStakeDetails is the current staking contract's
address.