Tests covering scenario of staking with multiple staking contracts
We need to make sure operator addresses can be reused between staking
contracts as well as that the source of the stake (grant vs liquid
tokens) is properly identified. In the last commit, there was a fix for
this particular scenario. TokenStaking was incorrectly identifying stake
as coming from a grant if there was a grant staked to the same operator
with a previous staking contract.