Allow to top-up a stake during initialization period
In case the stake is still in the initialization period, we add tokens
immetiatelly but we also reset the stake initialization time counter.

This change allows for two things:
- Staker no longer have to pay for two transactions to achieve the same
result (cancel + stake) and operator address uniqueness is no longer a
problem.
- If stake has been undelegated from multiple operators and is scattered
in the escrow, staker may perform a delegation and they top-up as many
times as needed, right after the delegation. Multiple transactions have
to be sent to the chain but staker does no longer need to suffer from
initialization period between the delegation and the first top-up.

Because out-of-gas problems during TokenStaking deployment I had to
sacrifice something. I decided to shorten one revert string and to
remove oldOwner from StakeOwnershipTransferred event. This value can be
deducted from previous StakeOwnershipTransferred event for the given
operator or from the Staked event.