Respect redelegated tokens when withdrawing revoked grant from escrow
TokenStakingEscrow was not respecting redelegated tokens in
withdrawRevoked function. As a result, grant manager could more tokens
from the escrow than they should be allowed to and even drain the escrow
to zero in the most pessimistic case.

The solution is to call availableAmount function to evaluate the amount
of tokens available for withdrawal from a revoked grant, just like we do
during the migration of tokens between escrows.