TokenStakingEscrow reverts if it already has a deposit
In case someone calls cancelStake twice they can have their information
about the deposit lost and tokens locked in the escrow. We prevent it by
reverting the transaction if deposit for the given operator is already
in the escrow. The same situation and the same fix is for calling
recoverStake twice. This change affects only grant stakes since liquid
token stakes are returned straight to the owner.