Download secrets
This is about as rudimentary as it gets but here's a thing to download
each secret from a cluster.

It expects you to be on the correct kube context and VPN for the
environment you want to download the secrets from. Key values are base64
encoded but the secret metadata should have the actuals.

Frankly, this is a messy way to work with secrets, but if you don't want
to maintain a clean / plaintext representation of the configurations in
a secure location, this is the way.