Do not allow to redelegate to an operator from which delegation has been canceled
After we added a feature allowing to top-up in the initialization period we
unconsciously introduced a path allowing to redelegate to the operator from which
the delegation has been canceled. It should not be allowed as it leads to rejecting
the deposit by the escrow when undelegating/canceling that second delegation.

Here we make sure that it is not possible to delegate to the operator from which
the delegation has been canceled.