Prevent top-ups on canceled delegations - improvements
There are two scenarios when top-up can be performed on a canceled
delegation:

1. Someone tries to re-delegate from the escrow to operator that already
has a deposit in the escrow:

delegate operator A -> escrow ->
redelegate operator B -> escrow ->
redelegate operator A -> escrow (TX reverted!)

This scenario is now prohibited from happening by
TokenStakingEscrow.hasDeposit check in redelegate function.

2. Someone tries to top-up a stake to an operator that has been
previously used and from which delegation was canceled.

This scenario is now prohibited from happening because of
TokenStakingEscrow.hasDeposit check in the TopUps.executeInOneStep
function.

The problem described here is not a vulnerability when a 3rd party can
hurt a staker but rather an interesting edge case when the staker can
hurt themselves.

When working on the fix, I had to move some code out of TokenStaking to
TopUps library to avoid out-of-gas issues. The good part about it is
that we now have the entire logic but grant rules validation in TopUps
library.