Refactored how a pending top-up is committed.
Instead of calculating new operator params in TokenStaking and emitting
event there, we can do all of it inside commit function, just like we do
when performing top-up in one step in executeInOneStep function.

I have also renamed forceCommit to cancel as it better reflects whats
happening - we are not modifying operatorParams there.

This change allows to align the commit code a little better with
initiateTopUp / executeInOneStep functions. It also allows to save some
bytecode in TokenStaking.