Include family with machine image
I guess when we updated the Google provider it changed the pathing for
machine image location.  The module defaults to ubuntu-1804-lts without
the family prefix.

Adding the prefix here addressed the issue, leaving the module  as-is
since we can override in the environment config.  May want to update
after folks migrate to the new Google provider.