Moving ropsten-miner to default namespace
For a bit we've been running the Ropsten miner on a dedicated node-pool
and namespace.  After scaling up the default node pool in keep-dev we
migrated the ropsten-miner back to the default namespace and had to
adjust naming not to collide with the internal ETH network components we
were running.