Minimum stake schedule starts at the moment of deploying KEEP token
We are deploying a new staking contract version but we want to maintain
minimum KEEP stake schedule. To make it happen, instead of using staking
contract deployment time as a schedule start, we now define a constant
in MinimumStakeSchedule with a value equal to the time of deploying KEEP
token. On the other hand, we want the slashing schedule (1% for the
first 3 month and 50% for the next 3 months) start with the time of
deploying TokenStaking contract. I renamed
TokenStaking.minimumStakeScheduleStart to TokenStaking.deployedAt and
updated Groups.sol to use the new name.

The most problematic part was tests. Since we now use a fixed point in
the past as a minimum stake schedule start, I had to introduce
TokenStakingStub for some tests with a fixed minimum stake of 100k KEEP.