Allow to copy stake only one time even if it's been recovered
Modified StakingPortBacker to allow for copying stake only one time even
if the copied stake has been recovered. This behaviour was expected
based on the documentation on copyStake which says:

"(...)  the relationship can only be copied once (...)".

This scenario would be the most problematic in the situation when
StakingPortBacker owner force-undelegates the stake after 3 months of
not paying the loan back and recovers the stake after the undelegation
time and the malicious staker copies the stake immediatelly before
StakingPortBacker owner has a chance to withdraw tokens (gas auction).