Added test covering recoverStake called twice
Added a unit test for not yet covered scenario when owner of a
delegation created from liquid tokens calls recoverStake twice and
making sure no more tokens than those from the original delegation are
transferred back.

For delegation created from a grant we have a test case in
TestStakingGrant making sure recoverStake fails when called for the
second time (stake is already in the escrow).