Dropped removeServiceContract from the operator contract
We remove this function to save some bytecode for gas price ceiling
updates. At the same time, there was a little to no benefit of having
this function. There are two places we guard by onlyServiceContract
modifier: function triggering group selection and function requesting
for a new entry signature. Both those functions are designed and
implemented in a way that is secure for staker and operator contract
state even if the service contract gets compromised.