Replaced ownerOperators mapping with an event
ownerOperators mapping in StakeDelegatable was quite a problematic
append-only data structure.

The first issue was that this mapping was not cleared up when a stake
was recovered. The reason for that was to maintain a historical mapping
between owner and operator so that the owner may withdraw rewards after
recovering the stake even if it lost the operator address (operator
address is a parameter to withdraw function).

Another issue was that this mapping maintained no longer up-to-date
information in case the stake ownership was transferred with
transferStakeOwnership function in TokenStaking. Clearing up this mapping
would involve writing a loop and compering operator address bytes. This
is problematic because of the contract size constraints as well as there
is a tiny possibility such a loop may cause gas limit problems in case an
owner has a lot of delegations.

The solution proposed here assumes we drop ownerOperators altogether,
introduce StakeDelegated event with indexed owner and operator, as well
as rename Stake event to OperatorStaked and remove owner field from there.

The relationship between owner and operator can be now looked up using
StakeDelegatedEvent having both owner and operator indexed. This event is
emitted when the delegation to the operator is registerred for the first
time as well as when the delegation is transferred to another owner.
OperatorStaked event contains information about operator-beneficiary-authorizer
triplet and this triplet is never changing. Since those values are set only
one time and never updated, we emit this event only one time - when the
delegation is created.

From KEEP token dashboard perspective:
- To find authorizer for operator, it's enough to filter out OperatorStaked
  event by indexed fields
- To find beneficiary for operator, it's enough to filter out OperatorStaked
  event by indexed fields
- To find operator for beneficiary, it's enough to filter out OperatorStaked
  event by indexed fields
- To find operator for authorizer, it's enough to filter out OperatorStaked
  event by indexed fields
- To find operator to owner, we should:
  Filter out StakeDelegated events by owner field
  Filter out StakeDelegated events by operator field, checking more recent
  events to make sure the delegation ownership has not been transferred.