Update event subscribers
Here we updated all `On*` functions of the
chain interface to not return an error. Instead
of that the error is logged inside the
implementation code. This is motivated by the
fact that `Watch*` functions never return the
error (despite they declare it) so we should not
allow to leak the problem out of the chain
interface.