Grant delegations only from grantee's perspective
To get grantee grants the dapp calls `getGrants` from `TokenGrant`
contract. The `getGrants` function returns grant ids for a grant manager
or a grantee. So it's possible that the proivded address is a grantee in
grant A and grant manager in grant B. In that case a `getGrants`
function retursn -> [A, B]. We need to filter out result from this
function and check if the provided address is a grantee in the provided
grant.