Fetching copied delegations
We should display the copied delegation on the tokens page. Currently
the owner can't `add keep`/`undelegate`/`recover` delegation. It will be
resolved in the future PR. Next to the copied delegation we display a
badge that says `this delegation is copied`.

To fetch copied delegations we should:
* scan `StakeCopied` event by owner field,
* scan `StakeDelegated` event by owner(StakingPortBacker address) and
operators from ^
* get details about delegation for each operator address.