Fix initcontainers that require keyfile
The provisioning script used in init containers recently moved to
reading Ethereum key and account information from a keyfile rather than
from the environment; however, the corresponding Kubernetes stateful set
and cron job configurations were not updated to mount the keyfile for
the initcontainer (it was already being mounted for the container
proper).

The initcontainer now has the keyfile mounted properly. In the process,
moved volumes to the top of the spec, since they are now used in both of
the subsequent sections.