Clone cached results from CacheSubprovider
The cache subprovider caches results, but the cached values are mutable,
and sure enough, the downstream handlers can and do at times mangle the
results in non-idempotent ways. This means that when they receive cached
values that they've already mangled later, everything blows up. This
mini-middleware clones the results at the two exit points that the cache
subprovider can use, ensuring that any downstream handlers are mutating
a request-specific version of the value, without mangling the cached
version.